Starting Refactor
