+++ Cargo.toml
@@ -20,6 +20,9 @@ reqwest = { version = "0.11.4", features = ["json"] }
 serde = { version = "1.0.114", features = ["derive"] }
 serde_json = "1.0.57"
 regex = "1.5"
+roxmltree = "0.19"
+html-escape = "0.2"
+log = "0.4.20"
 
 [dev-dependencies]
