+++ src/client/mod.rs
@@ -1,0 +1,3 @@
+pub mod client;
+pub mod bulk_api;
+pub mod bulk_api_v2;
