+++ src/responses/create_response.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct CreateResponse {
+    pub id: String,
+    pub success: bool,
+}
