+++ src/responses/mod.rs
@@ -1,0 +1,18 @@
+pub mod create_response;
+pub mod upsert_response;
+pub mod token_response;
+pub mod query_response;
+pub mod error_response;
+pub mod register_response;
+pub mod token_error_response;
+pub mod describe_response;
+pub mod bulk_api_v1_response;
+pub mod bulk_api_v2_response;
+pub mod bulk_api_v2_state_change_response;
+pub mod bulk_api_v2_status_response;
+pub mod describe_global_response;
+pub mod describe_global_sobject_response;
+pub mod search_response;
+pub mod version_response;
+pub mod all_jobs_response;
+pub mod job_status_response;
