+++ src/responses/token_error_response.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct TokenErrorResponse {
+    error: String,
+    error_description: String,
+}
