+++ src/responses/version_response.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct VersionResponse {
+    pub label: String,
+    pub url: String,
+    pub version: String,
+}
